test(RestaurantControllerTest)

+++ src/main/java/com/kcjcustomerbe/service/impl/ReviewServiceImpl.java
@@ -5,14 +5,12 @@ import com.kcjcustomerbe.entity.Customer;
 import com.kcjcustomerbe.entity.Restaurant;
 import com.kcjcustomerbe.entity.Review;
 import com.kcjcustomerbe.exception.ErrorMessage;
-import com.kcjcustomerbe.exception.list.customer.CustomerNotFoundException;
-import com.kcjcustomerbe.exception.list.IdNotFoundException;
-import com.kcjcustomerbe.exception.list.restaurant.RestaurantNotFoundException;
-import com.kcjcustomerbe.exception.list.review.ReviewEmptyException;
+import com.kcjcustomerbe.exception.list.*;
 import com.kcjcustomerbe.mapper.ReviewMapper;
 import com.kcjcustomerbe.repo.CustomerRepository;
 import com.kcjcustomerbe.repo.RestaurantRepository;
 import com.kcjcustomerbe.repo.ReviewRepository;
+import com.kcjcustomerbe.service.RestaurantService;
 import com.kcjcustomerbe.service.ReviewService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -34,7 +32,7 @@ public class ReviewServiceImpl implements ReviewService {
 
    // CREATE
    @Override
-   public ReviewDto addReview(ReviewDto reviewDto, UUID customerId, Long restaurantId) throws IdNotFoundException {
+   public ReviewDto addReview(ReviewDto reviewDto, UUID customerId, Long restaurantId) {
       Customer customer = customerRepository.findById(customerId)
             .orElseThrow(() -> new CustomerNotFoundException(ErrorMessage.CUSTOMER_ID_NOT_FOUND + customerId));
 
@@ -42,18 +40,17 @@ public class ReviewServiceImpl implements ReviewService {
             .orElseThrow(() -> new RestaurantNotFoundException(ErrorMessage.RESTAURANT_ID_NOT_FOUND + restaurantId));
 
       if (reviewDto != null) {
-      Review review = reviewMapper.mapReviewToCreateReviewDto(reviewDto);
+         Review review = reviewMapper.mapReviewToCreateReviewDto(reviewDto);
 
-      review.setCustomer(customer);
-      review.setRestaurant(restaurant);
+         review.setCustomer(customer);
+         review.setRestaurant(restaurant);
 
-      Review savedReview = reviewRepository.save(review);
-      return reviewMapper.mapReviewToReviewDto(savedReview);
+         Review savedReview = reviewRepository.save(review);
+         return reviewMapper.mapReviewToReviewDto(savedReview);
 
       } else {
          throw new ReviewEmptyException(ErrorMessage.REVIEW_BODY_IS_EMPTY);
       }
-
    }
 
    // READ
@@ -71,7 +68,7 @@ public class ReviewServiceImpl implements ReviewService {
       List<Review> reviews = reviewRepository.findAll();
 
       if (!reviews.isEmpty()) {
-      return reviewMapper.mapReviewsToReviewsDto(reviews);
+         return reviewMapper.mapReviewsToReviewsDto(reviews);
       } else {
          throw new ReviewEmptyException(ErrorMessage.REVIEWS_NOT_FOUND);
       }
@@ -88,6 +85,11 @@ public class ReviewServiceImpl implements ReviewService {
    // DELETE
    @Override
    public void deleteReview(Long reviewId) {
+
+      if (reviewId == null) {
+         throw new ReviewNotFoundException(null);
+      }
+
       reviewRepository.deleteById(reviewId);
    }
 }
